requirements.txt
flask==3.0.0
requests>=2.31.0
beautifulsoup4==4.12.2
openai==1.12.0
python-dotenv==1.0.0
apscheduler==3.10.4
pykrx>=1.2.0
